+++ src/lib/components/modals/formModals/FormModalExercise.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { exerciseSchema } from '$lib/exercise';
+  import { exerciseSchema } from '$lib/zodSchemas';
   import FormBodyExercise from '../../forms/bodies/FormBodyExercise.svelte';
   import FormModal from './FormModal.svelte';
 </script>
