+++ src/lib/zodSchemas.ts
@@ -1,5 +1,14 @@
 import { evaluate } from 'mathjs';
 import { z } from 'zod';
+import {
+  difficulties,
+  equipments,
+  exerciseEventFieldsToShow,
+  exerciseTypes,
+  muscleGroups,
+  muscles,
+  postures,
+} from './utils';
 
 export const exerciseEventSchema = z.object({
   date: z.date().default(new Date()).nullish(),
@@ -101,3 +110,20 @@ export const customQueryConditionSchema = z
     }
   });
 export type CustomQueryConditionSchema = typeof customQueryConditionSchema;
+
+export const exerciseSchema = z.object({
+  name: z.string().min(1),
+  type: z.enum(exerciseTypes),
+  difficulty: z.enum(difficulties).nullish(),
+  videoUrl: z.string().nullish(),
+  muscleGroup: z.enum(muscleGroups).nullish().default(null),
+  primeMoverMuscle: z.enum(muscles).nullish().default(null),
+  secondaryMuscle: z.enum(muscles).nullish().default(null),
+  tertiaryMuscle: z.enum(muscles).nullish().default(null),
+  primaryEquipment: z.enum(equipments).nullish().default(null),
+  posture: z.enum(postures).nullish().default(null),
+  fieldsToShow: z.array(z.enum(exerciseEventFieldsToShow)).refine((val) => val.length > 0, {
+    message: 'At least one field is required',
+  }),
+});
+export type ExerciseSchema = typeof exerciseSchema;
